// resources/js/types/InventariablesFormValues.ts
export type InventariablesFormValues = {
    tipo: string;
    clasificacion_id: number;
    alta: string;
    grupo_activo: string;
    area_id: number | null;
    fecha_ingreso: string;
    numero_inventario: string;
    depreciacion: string;
    nombre: string;
    marca: string;
    modelo: string;
    serie: string;
    numero_motor: string;
    numero_factura: string;
    proveedor_id: number;
    estado: string;
    costo_unitario: string;
    numero_economico: string;
    placas: string;
    tipo_poliza: string;
    numero_poliza: string;
    descripcion: string;
    imagenes: FileList;
    estatus: string;
}